Refactor command tests into one test file.
