--- tests/AnalyzeTest.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * 
- *
- * @author Sam Schmidt
- * @date 2015-05-04
- * @company Linus Shops
- */
-
-namespace LinusShops\Prophet;
-
-class AnalyzeTest extends \PHPUnit_Framework_TestCase
-{
-    public function testExample()
-    {
-
-    }
-}
